trading_system/requirements.txt
MetaTrader5
pandas
python-dotenv
pytz
requests
